LayoutTests/http/tests/paymentrequest/rejects_if_not_active.https-expected.txt


PASS PaymentRequest.show() aborts if the document is not active
PASS PaymentRequest.show() aborts if the document is active, but not fully active
PASS If a payment request is showing, but its document is navigated away (so no longer fully active), the payment request aborts.
